Fix some linting problems.
- `static_assert` with no message is only introduced in C++17 (and is not good)
- Removed unused `context` variable
- Changed whenever we were multiplying by 8 (number of bits in 1 byte) by multiplication
  by `CHAR_BIT` (in `limits.h`)

+++ include/tdnat/jit_api.h
@@ -79,7 +79,7 @@ struct OptionalArrayRef {
 
   static c10::OptionalArrayRef<T> *run(const T *ptr, int64_t size)
   {
-    return new c10::OptionalArrayRef<T>({ptr, size});
+    return new c10::OptionalArrayRef<T>({ptr, static_cast<size_t>(size)});
   }
 };
 
@@ -103,10 +103,10 @@ struct List {
     return "List<" + LLVMType<T>::name() + ">";
   }
 
-  static c10::List<T> *run(T *list, size_t size)
+  static c10::List<T> *run(T *list, int64_t size)
   {
     // Constructor is marked as explicit.
-    return new c10::List<T>({list, size});
+    return new c10::List<T>({list, static_cast<size_t>(size)});
   }
 };
